+++ packages/didcomm/src/DidCommEnvelopeService.ts
@@ -16,7 +16,7 @@ import { getResolvedDidcommServiceWithSigningKeyId } from './modules/connections
 import { DidCommOutOfBandRole } from './modules/oob/domain/DidCommOutOfBandRole'
 import { DidCommOutOfBandRepository } from './modules/oob/repository/DidCommOutOfBandRepository'
 import { DidCommOutOfBandRecordMetadataKeys } from './modules/oob/repository/outOfBandRecordMetadataTypes'
-import { DidCommForwardMessage } from './modules/routing/messages/DidCommForwardMessage'
+import { DidCommForwardMessage } from './modules/routing/messages/v1/DidCommForwardMessage'
 import { DidCommMediatorRoutingRepository } from './modules/routing/repository/DidCommMediatorRoutingRepository'
 import { DidCommDocumentService } from './services/DidCommDocumentService'
 import type { DidCommEncryptedMessage, DidCommPlaintextMessage } from './types'
