+++ packages/didcomm/src/modules/routing/DidCommRoutingEvents.ts
@@ -1,6 +1,6 @@
 import type { BaseEvent } from '@credo-ts/core'
 import type { DidCommRouting } from '../../models'
-import type { DidCommKeylistUpdate } from './messages/DidCommKeylistUpdateMessage'
+import type { DidCommKeylistUpdate } from './messages/v1/DidCommKeylistUpdateMessage'
 import type { DidCommMediationState } from './models/DidCommMediationState'
 import type { DidCommMediationRecord } from './repository/DidCommMediationRecord'
 import type { KeylistUpdateResponseItem } from './messages/v2/DidCommKeylistUpdateResponseV2Message'
