+++ packages/didcomm/src/modules/routing/handlers/index.ts
@@ -1,7 +1,2 @@
-export * from './DidCommForwardHandler'
-export * from './DidCommKeylistUpdateHandler'
-export * from './DidCommKeylistUpdateResponseHandler'
-export * from './DidCommMediationDenyHandler'
-export * from './DidCommMediationGrantHandler'
-export * from './DidCommMediationRequestHandler'
+export * from './v1'
 export * from './v2'
