+++ packages/didcomm/src/modules/routing/handlers/v1/DidCommKeylistUpdateResponseHandler.ts
@@ -1,6 +1,6 @@
-import type { DidCommMessageHandler, DidCommMessageHandlerInboundMessage } from '../../../handlers'
-import { DidCommKeylistUpdateResponseMessage } from '../messages'
-import type { DidCommMediationRecipientService } from '../services'
+import type { DidCommMessageHandler, DidCommMessageHandlerInboundMessage } from '../../../../handlers'
+import { DidCommKeylistUpdateResponseMessage } from '../../messages'
+import type { DidCommMediationRecipientService } from '../../services'
 
 export class DidCommKeylistUpdateResponseHandler implements DidCommMessageHandler {
   public mediationRecipientService: DidCommMediationRecipientService
