+++ packages/didcomm/src/modules/routing/handlers/v1/index.ts
@@ -1,0 +1,6 @@
+export * from './DidCommForwardHandler'
+export * from './DidCommKeylistUpdateHandler'
+export * from './DidCommKeylistUpdateResponseHandler'
+export * from './DidCommMediationDenyHandler'
+export * from './DidCommMediationGrantHandler'
+export * from './DidCommMediationRequestHandler'
