+++ packages/didcomm/src/modules/routing/messages/index.ts
@@ -1,7 +1,2 @@
-export * from './DidCommForwardMessage'
-export * from './DidCommKeylistUpdateMessage'
-export * from './DidCommKeylistUpdateResponseMessage'
-export * from './DidCommMediationDenyMessage'
-export * from './DidCommMediationGrantMessage'
-export * from './DidCommMediationRequestMessage'
+export * from './v1'
 export * from './v2'
