+++ packages/didcomm/src/modules/routing/messages/v1/DidCommMediationGrantMessage.ts
@@ -1,9 +1,9 @@
 import { Expose } from 'class-transformer'
 import { IsArray, IsNotEmpty, IsString } from 'class-validator'
 
-import { DidCommMessage } from '../../../DidCommMessage'
-import { IsValidMessageType, parseMessageType } from '../../../util/messageType'
-import type { DidCommVersion } from '../../../util/didcommVersion'
+import { DidCommMessage } from '../../../../DidCommMessage'
+import { IsValidMessageType, parseMessageType } from '../../../../util/messageType'
+import type { DidCommVersion } from '../../../../util/didcommVersion'
 
 export interface DidCommMediationGrantMessageOptions {
   id?: string
