+++ packages/didcomm/src/modules/routing/messages/v1/index.ts
@@ -1,0 +1,7 @@
+export * from './DidCommForwardMessage'
+export * from './DidCommKeylistMessage'
+export * from './DidCommKeylistUpdateMessage'
+export * from './DidCommKeylistUpdateResponseMessage'
+export * from './DidCommMediationDenyMessage'
+export * from './DidCommMediationGrantMessage'
+export * from './DidCommMediationRequestMessage'
